+++ src/inspect_ai/_display/rich.py
@@ -1,6 +1,7 @@
 import asyncio
 import contextlib
 import datetime
+from contextlib import contextmanager
 from dataclasses import dataclass
 from typing import Any, Callable, Iterator, Set
 
@@ -15,6 +16,7 @@ from rich.progress import (
     TimeElapsedColumn,
 )
 from rich.progress import Progress as RProgress
+from rich.segment import Segment
 from rich.table import Table
 from rich.text import Text
 from typing_extensions import override
@@ -25,6 +27,7 @@ from inspect_ai._util.platform import is_running_in_jupyterlab, is_running_in_vs
 from inspect_ai._util.throttle import throttle
 from inspect_ai.log import EvalStats
 from inspect_ai.log._log import rich_traceback
+from inspect_ai.log._transcript import InputEvent, transcript
 from inspect_ai.util._concurrency import concurrency_status
 
 from ._display import (
@@ -181,6 +184,9 @@ class RichTaskScreen(TaskScreen):
         # show cursor for input
         self.live.console.show_cursor(True)
 
+        # record console activity for event
+        self.live.console.record = True
+
         try:
             # print header if requested
             if header:
@@ -189,11 +195,18 @@ class RichTaskScreen(TaskScreen):
                 self.live.console.print("")
 
             # yield the console
-            yield self.live.console
+            with record_console_input(self.live.console):
+                yield self.live.console
 
             # print one blank line
             self.live.console.print("")
         finally:
+            # capture recording then yield input event
+            input = self.live.console.export_text(clear=False, styles=False)
+            input_ansi = self.live.console.export_text(clear=True, styles=True)
+            self.live.console.record = False
+            transcript()._event(InputEvent(input=input, input_ansi=input_ansi))
+
             # disable cursor while not collecting input
             self.live.console.show_cursor(False)
 
@@ -675,3 +688,21 @@ def rich_progress() -> RProgress:
 
 _theme: Theme | None = None
 _display: RichDisplay | None = None
+
+
+@contextmanager
+def record_console_input(console: Console) -> Iterator[None]:
+    input_original = Console.input
+
+    def input_with_record(self: Console, *args: Any, **kwargs: Any) -> str:
+        result = input_original(self, *args, **kwargs)
+        if self.record:
+            with self._record_buffer_lock:
+                self._record_buffer.append(Segment(result))
+        return result
+
+    Console.input = input_with_record  # type: ignore
+
+    yield
+
+    Console.input = input_original  # type: ignore